Migrate shared-folder models (tabicl, tabpfnv2_5) to per-model layout
These folders host multiple AutoGluon model classes:
- tabicl/: TabICLModel + TabICLv2Model (only v2 has dedicated MethodMetadata)
- tabpfnv2_5/: RealTabPFNv25Model + TabPFNv26Model

`hpo.py` holds the shared search-space machinery and exports both `gen_*`
objects. `info.py` declares one `ModelInfo` per model class with a dedicated
MethodMetadata entry; auto-discovery picks them up by `ag_name`.

The dated metadata files (_2026_02_16, _2025_11_12, _2026_03_18) now source
`tabiclv2_metadata`, `realtabpfn25_metadata`, and `tabpfn26_metadata` from
each model's `info.py`. Legacy aliases preserved.
