+++ tabarena/tabarena/models/tabicl/__init__.py
@@ -1,0 +1,11 @@
+from __future__ import annotations
+
+from tabarena.models.tabicl.hpo import gen_tabicl, gen_tabiclv2
+from tabarena.models.tabicl.info import tabiclv2_info, tabiclv2_method_metadata
+
+__all__ = [
+    "gen_tabicl",
+    "gen_tabiclv2",
+    "tabiclv2_info",
+    "tabiclv2_method_metadata",
+]
