+++ tabarena/tabarena/models/tabpfnv2_5/__init__.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+from tabarena.models.tabpfnv2_5.hpo import gen_realtabpfnv25, gen_tabpfnv26
+from tabarena.models.tabpfnv2_5.info import (
+    realtabpfnv25_info,
+    realtabpfnv25_method_metadata,
+    tabpfnv26_info,
+    tabpfnv26_method_metadata,
+)
+
+__all__ = [
+    "gen_realtabpfnv25",
+    "gen_tabpfnv26",
+    "realtabpfnv25_info",
+    "realtabpfnv25_method_metadata",
+    "tabpfnv26_info",
+    "tabpfnv26_method_metadata",
+]
